Forwards errors from python to scripts

+++ backup.sh
@@ -16,4 +16,10 @@ source "$directory/venv/bin/activate"
 
 python3 $directory/main.py $2 $3 $4 $5 $temp_file
 
+# If the python script has errored, stop start.sh here
+if [ $? -ne 0 ]; then
+    exit 1
+fi
+
+
 sudo rm $temp_file
