+++ scripts/mongobackup/mongo_backup.sh
@@ -17,5 +17,10 @@ source "$directory/venv/bin/activate"
 
 python3 $directory/main.py $2 $3 $4 $5 "$temp_folder.tar"
 
+# If the python script has errored, stop start.sh here
+if [ $? -ne 0 ]; then
+    exit 1
+fi
+
 sudo rm -rf $temp_folder
 sudo rm "$temp_folder.tar"
